[feat/#64]: 탑랭킹 레이아웃 수정
